pulls : move around pr list stuff
